requirements.txt
streamlit==1.38.0
python-docx==0.8.11
docxcompose==1.4.0
